gui-vision/requirements.txt
openai>=1.0.0
mss>=9.0.0
pyautogui>=0.9.54
Pillow>=10.0.0
PyYAML>=6.0
pyperclip>=1.8.0